Removed old grpc and renamed grpc1 to grpc where the new chat service exists.
