+++ services/grpc/.devcontainer/docker-compose.yml
@@ -1,9 +1,9 @@
 services:
-  grpc-service:
-    container_name: grpc-service
+  grpc-chat-service:
+    container_name: grpc-chat-service
     image: mcr.microsoft.com/devcontainers/python:3.12
     volumes:
       - ../../..:/workspace:cached
     ports:
-      - "50051:50051"
+      - "50052:50052"
     command: sleep infinity
