+++ services/grpc/README.md
@@ -1,6 +1,6 @@
-# gRPC Library Service
+# gRPC Chat Service
 
-A gRPC service for the library domain — fetch books, create new ones, and subscribe to live updates when books are added. Built with Python and grpcio, using SQLite as the database.
+A gRPC chat service — send messages and join a live bidirectional chat stream. Built with Python and grpcio, using SQLite as the database.
 
 ---
 
@@ -10,7 +10,7 @@ A gRPC service for the library domain — fetch books, create new ones, and subs
 
 - Uses **Protocol Buffers (protobuf)** for compact binary serialization — smaller and faster than JSON/XML
 - Defines its API contract in a **`.proto` file**, which is used to generate server and client code in any language
-- Supports **streaming** — the server can push multiple messages to the client over a single connection
+- Supports **streaming** — both client and server can send multiple messages over a single connection
 - Runs on top of **HTTP/2**, enabling multiplexed connections
 
 ### Key concepts
@@ -19,7 +19,7 @@ A gRPC service for the library domain — fetch books, create new ones, and subs
 |------|--------------|
 | **`.proto` file** | Defines the service contract — messages (data shapes) and RPCs (operations). Both server and client code are generated from this. |
 | **Unary RPC** | Standard request-response — client sends one message, server replies with one message. |
-| **Server streaming RPC** | Client sends one request, server sends back a stream of messages over time. |
+| **Bidirectional streaming RPC** | Both client and server send streams of messages simultaneously over a single connection. |
 | **Protobuf message** | A typed data structure defined in the `.proto` file. Serialized to compact binary on the wire. |
 | **Status codes** | gRPC uses its own status codes (`NOT_FOUND`, `INVALID_ARGUMENT`, etc.) instead of HTTP status codes. |
 
@@ -39,23 +39,23 @@ A gRPC service for the library domain — fetch books, create new ones, and subs
 ## Project Structure
 
 ```
-grpc/
+grpc1/
 ├── .devcontainer/
 │   ├── devcontainer.json     # VS Code dev container config
 │   └── docker-compose.yml    # Container definition
 ├── proto/
-│   └── library.proto         # Service contract — defines RPCs and messages
+│   └── chat.proto            # Service contract — defines RPCs and messages
 ├── src/
 │   ├── server.py             # Entry point — starts the gRPC server
 │   ├── services/
-│   │   └── library_service.py # RPC handler implementations
+│   │   └── chat_service.py   # RPC handler implementations
 │   ├── db/
-│   │   └── database.py       # SQLite connection helper
+│   │   └── database.py       # SQLite connection helper + table init
 │   └── generated/            # Auto-generated protobuf/gRPC code (do not edit)
-│       ├── library_pb2.py
-│       └── library_pb2_grpc.py
+│       ├── chat_pb2.py
+│       └── chat_pb2_grpc.py
 ├── data/
-│   └── library.db            # SQLite database
+│   └── chat.db               # SQLite database (created at runtime)
 ├── Dockerfile
 ├── requirements.txt
 └── README.md
@@ -71,24 +71,24 @@ grpc/
 
 ### Run with Dev Container
 
-1. Open VS Code and run **Dev Containers: Reopen in Container** (select the gRPC Service container).
+1. Open VS Code and run **Dev Containers: Reopen in Container** (select the gRPC Chat Service container).
 
-2. Dependencies are installed automatically via `postCreateCommand`.
+2. Dependencies are installed automatically via `postCreateCommand`, which also runs the protoc compiler.
 
-3. Generate protobuf code (only needed once, or after changing the `.proto` file):
+3. If you need to regenerate protobuf code manually (after changing the `.proto` file):
 
 ```bash
 python -m grpc_tools.protoc \
   -I./proto \
   --python_out=./src/generated \
   --grpc_python_out=./src/generated \
-  ./proto/library.proto
+  ./proto/chat.proto
 ```
 
 4. Fix the generated import (protoc generates flat imports):
 
 ```bash
-sed -i 's/import library_pb2/from src.generated import library_pb2/' ./src/generated/library_pb2_grpc.py
+sed -i 's/import chat_pb2/from src.generated import chat_pb2/' ./src/generated/chat_pb2_grpc.py
 ```
 
 5. Start the server:
@@ -97,7 +97,7 @@ sed -i 's/import library_pb2/from src.generated import library_pb2/' ./src/gener
 python -m src.server
 ```
 
-Server listens on `0.0.0.0:50051`.
+Server listens on `0.0.0.0:50052`.
 
 ---
 
@@ -105,49 +105,41 @@ Server listens on `0.0.0.0:50051`.
 
 | RPC | Type | Description |
 |-----|------|-------------|
-| `GetBookById` | Unary | Fetch a book by its ID. Returns `NOT_FOUND` if it doesn't exist. |
-| `CreateBook` | Unary | Create a new book. Validates author/publisher exist and year >= 1900. Returns the generated ID. |
-| `WatchBooks` | Server streaming | Subscribe to live updates — receives a `Book` message every time any client creates a new book. |
+| `SendMessage` | Unary | Store a single chat message. Validates non-empty username and content. Returns the generated ID and timestamp. |
+| `Chat` | Bidirectional streaming | Client streams messages in; server broadcasts every message to **all** connected clients in real-time. |
 
-### GetBookById
-
-```json
-// Request
-{ "book_id": 1000 }
-
-// Response
-{
-  "book_id": 1000,
-  "title": "Harry Potter and the Sorcerer's Stone (Book 1)",
-  "author_id": 394,
-  "publishing_company_id": 74,
-  "publishing_year": 1997
-}
-```
-
-### CreateBook
+### SendMessage
 
 ```json
 // Request
 {
-  "title": "Test Book",
-  "author_id": 1,
-  "publishing_company_id": 1,
-  "publishing_year": 2024
+  "username": "alice",
+  "content": "Hello everyone!"
 }
 
 // Response
-{ "book_id": 2002 }
+{
+  "message_id": 1,
+  "timestamp": "2026-05-20T10:00:00Z"
+}
 ```
 
-### WatchBooks
+### Chat
 
 ```json
-// Request
-{}
+// Each message sent by the client (stream)
+{
+  "username": "alice",
+  "content": "Hey everyone!"
+}
 
-// Response (streamed — one message per new book created by any client)
-{ "book_id": 2002, "title": "Test Book", "author_id": 1, ... }
+// Response stream — one message broadcast per message sent by any connected client
+{
+  "message_id": 1,
+  "username": "alice",
+  "content": "Hey everyone!",
+  "timestamp": "2026-05-20T10:00:00Z"
+}
 ```
 
 ---
@@ -158,8 +150,7 @@ gRPC uses **status codes** instead of HTTP status codes:
 
 | Status Code | When it is returned |
 |-------------|-------------------|
-| `NOT_FOUND` | Book ID does not exist (GetBookById) |
-| `INVALID_ARGUMENT` | Missing/invalid fields, nonexistent author or publisher, year < 1900 (CreateBook) |
+| `INVALID_ARGUMENT` | Empty username or empty content (SendMessage) |
 
 ---
 
@@ -168,10 +159,17 @@ gRPC uses **status codes** instead of HTTP status codes:
 ### Postman
 
 1. Open Postman → **New** → **gRPC**
-2. Server URL: `localhost:50051`
-3. Import `proto/library.proto`
+2. Server URL: `localhost:50052`
+3. Import `proto/chat.proto`
 4. Select an RPC and send a request
 
-### Database
+### Database Schema
 
-The SQLite database contains 998 books (IDs 1000–1999), authors, and publishing companies.
+```sql
+CREATE TABLE tmessage (
+    nMessageID INTEGER PRIMARY KEY AUTOINCREMENT,
+    cUsername  TEXT NOT NULL,
+    cContent   TEXT NOT NULL,
+    cTimestamp TEXT NOT NULL DEFAULT (strftime('%Y-%m-%dT%H:%M:%SZ', 'now'))
+);
+```
