--- services/grpc/bruno/gRPC Tests/.gitignore
@@ -1,9 +1,0 @@
-# Secrets
-.env*
-
-# Dependencies
-node_modules
-
-# OS files
-.DS_Store
-Thumbs.db
