--- services/grpc/bruno/gRPC Tests/CreateBook.yml
@@ -1,17 +1,0 @@
-info:
-  name: CreateBook
-  type: grpc
-  seq: 2
-
-grpc:
-  url: localhost:50051
-  method: /library.LibraryService/CreateBook
-  methodType: unary
-  message: |-
-    {
-      "title": "Test Book10",
-      "author_id": 1,
-      "publishing_company_id": 1,
-      "publishing_year": 2004
-    }
-  auth: inherit
