--- services/grpc/bruno/gRPC Tests/GetBookById.yml
@@ -1,14 +1,0 @@
-info:
-  name: GetBookById
-  type: grpc
-  seq: 1
-
-grpc:
-  url: localhost:50051
-  method: /library.LibraryService/GetBookById
-  methodType: unary
-  message: |-
-    {
-      "book_id": 1000
-    }
-  auth: inherit
