--- services/grpc/bruno/gRPC Tests/WatchBooks.yml
@@ -1,11 +1,0 @@
-info:
-  name: WatchBooks
-  type: grpc
-  seq: 3
-
-grpc:
-  url: localhost:50051
-  method: /library.LibraryService/WatchBooks
-  methodType: server-streaming
-  message: "{}"
-  auth: inherit
