--- services/grpc/bruno/gRPC Tests/opencollection.yml
@@ -1,25 +1,0 @@
-opencollection: 1.0.0
-
-info:
-  name: gRPC Tests
-config:
-  protobuf:
-    protoFiles:
-      - type: file
-        path: ../../proto/library.proto
-  proxy:
-    inherit: true
-    config:
-      protocol: http
-      hostname: ""
-      port: ""
-      auth:
-        username: ""
-        password: ""
-      bypassProxy: ""
-bundled: false
-extensions:
-  bruno:
-    ignore:
-      - node_modules
-      - .git
