--- services/grpc1/.devcontainer/devcontainer.json
@@ -1,29 +1,0 @@
-{
-    "name": "gRPC Chat Service",
-    "dockerComposeFile": "docker-compose.yml",
-    "service": "grpc-chat-service",
-    "workspaceFolder": "/workspace/services/grpc1",
-    "customizations": {
-        "vscode": {
-            "extensions": [
-                "ms-python.python",
-                "ms-python.vscode-pylance",
-                "zxh404.vscode-proto3",
-                "aaron-bond.better-comments",
-                "usernamehw.errorlens"
-            ],
-            "settings": {
-                "python.defaultInterpreterPath": "/usr/local/bin/python",
-                "editor.formatOnSave": true
-            }
-        }
-    },
-    "features": {
-        "ghcr.io/devcontainers/features/common-utils": {
-            "installZsh": true,
-            "installOhMyZsh": true,
-            "configureZshAsDefaultShell": true
-        }
-    },
-    "postCreateCommand": "pip install -r requirements.txt && python -m grpc_tools.protoc -I./proto --python_out=./src/generated --grpc_python_out=./src/generated ./proto/chat.proto && sed -i 's/import chat_pb2/from src.generated import chat_pb2/' ./src/generated/chat_pb2_grpc.py"
-}
