--- services/grpc1/requirements.txt
@@ -1,4 +1,0 @@
-grpcio==1.68.1
-grpcio-tools==1.68.1
-grpcio-reflection==1.68.1
-protobuf==5.29.2
